resturctured vault directory by moving each client implementation close to their domain
