+++ crates/bitwarden-vault/src/cipher/attachment_client.rs
