+++ crates/bitwarden-vault/src/cipher/cipher_client.rs
