+++ crates/bitwarden-vault/src/cipher/mod.rs
@@ -1,7 +1,9 @@
 pub(crate) mod attachment;
+pub(crate) mod attachment_client;
 pub(crate) mod card;
 #[allow(clippy::module_inception)]
 pub(crate) mod cipher;
+pub(crate) mod cipher_client;
 pub(crate) mod cipher_permissions;
 pub(crate) mod field;
 pub(crate) mod identity;
@@ -14,11 +16,13 @@ pub(crate) mod ssh_key;
 pub use attachment::{
     Attachment, AttachmentEncryptResult, AttachmentFile, AttachmentFileView, AttachmentView,
 };
+pub use attachment_client::{ClientAttachments, DecryptFileError, EncryptFileError};
 pub use card::{CardBrand, CardView};
 pub use cipher::{
     Cipher, CipherError, CipherListView, CipherListViewType, CipherRepromptType, CipherType,
     CipherView,
 };
+pub use cipher_client::ClientCiphers;
 pub use field::FieldView;
 pub use identity::IdentityView;
 pub use login::{
