+++ crates/bitwarden-vault/src/collection_client.rs
