+++ crates/bitwarden-vault/src/folder_client.rs
