--- crates/bitwarden-vault/src/mobile/mod.rs
@@ -1,5 +1,0 @@
-pub(crate) mod attachment_client;
-pub(crate) mod cipher_client;
-pub(crate) mod collection_client;
-pub(crate) mod folder_client;
-pub(crate) mod password_history_client;
