+++ crates/bitwarden-vault/src/password_history_client.rs
